messaging/consumer/requirements.txt
google-cloud-pubsub==2.7.0
pytz==2021.1
